Add files via upload
This is the working code to this point. This will control the movement, main power, camera and night vision led. Still need to finish the html and css to get all buttons to appear on the screen, add servo control for the head and output battery voltage to browser page